+++ src/lib/stores/useMetadataList.ts
@@ -7,7 +7,7 @@ import type { QueryKey } from '@tanstack/svelte-query';
 import type { EventPacket, RxNostr } from 'rx-nostr';
 import { pipe } from 'rxjs';
 
-import { filterMetadataList, latestEachPubkey, scanArray } from './operators.js';
+import { filterMetadataList, scanLatestEachPubkey } from './operators.js';
 import type { ReqResult, RxReqBase } from './types.js';
 import { useReq } from './useReq.js';
 
@@ -19,6 +19,6 @@ export function useMetadataList(
 ): ReqResult<EventPacket[]> {
   // TODO: Add npub support
   const filters = [{ kinds: [0], authors: pubkeys, limit: pubkeys.length }];
-  const operator = pipe(filterMetadataList(pubkeys), latestEachPubkey(), scanArray());
+  const operator = pipe(filterMetadataList(pubkeys), scanLatestEachPubkey());
   return useReq({ rxNostr, queryKey, filters, operator, req, initData: [] });
 }